feat(graphql): add index directive rds support
