+++ packages/amplify-graphql-index-transformer/src/index.ts
@@ -1,3 +1,3 @@
 export { IndexTransformer } from './graphql-index-transformer';
 export { PrimaryKeyTransformer } from './graphql-primary-key-transformer';
-export { attributeTypeFromType } from './resolvers';
+export { attributeTypeFromType } from './resolvers/common';
