+++ packages/amplify-graphql-index-transformer/src/resolvers/index.ts
@@ -1,0 +1,2 @@
+export * from './query';
+export * from './resolvers';
